requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
app==0.0.1
asyncpg==0.30.0
attrs==24.3.0
Authlib==1.6.1
banks==2.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.115.6
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.2.4
griffe==1.11.0
grpcio==1.74.0
grpcio-health-checking==1.74.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langsmith==0.2.3
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.1
llama-index-cli==0.5.0
llama-index-core==0.13.1
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.1
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.2
llama-index-readers-file==0.5.0
llama-index-readers-llama-parse==0.5.0
llama-index-vector-stores-pinecone==0.7.0
llama-index-vector-stores-weaviate==1.4.0
llama-index-workflows==1.3.0
llama-parse
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.18.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.0
openai==1.99.6
openpyxl==3.1.5
orjson==3.10.12
packaging==24.2
pandas==2.2.3
passlib==1.7.4
password==0.2
pillow==11.0.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.11.1
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.36
starlette==0.41.3
streamlit==1.41.1
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.11.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.35.0
weaviate==0.1.2
weaviate-client==4.16.6
Werkzeug==3.0.4
wrapt==1.17.2
yarl==1.20.1